+++ src/server/game/Entities/Vehicle/Vehicle.h
@@ -96,4 +96,15 @@ private:
     Status _status;
 };
 
+class VehicleDespawnEvent : public BasicEvent
+{
+public:
+    VehicleDespawnEvent(Unit& self, uint32 duration) : _self(self), _duration(duration) { }
+    bool Execute(uint64 e_time, uint32 p_time) override;
+
+protected:
+    Unit& _self;
+    uint32 _duration;
+};
+
 #endif
